load_json now loads SimulationContainer; simulation refactor
